+++ cmd/tempo/app/modules.go
@@ -10,13 +10,14 @@ import (
 	cortex_transport "github.com/cortexproject/cortex/pkg/frontend/transport"
 	cortex_frontend_v1pb "github.com/cortexproject/cortex/pkg/frontend/v1/frontendv1pb"
 	"github.com/cortexproject/cortex/pkg/ring"
-	"github.com/cortexproject/cortex/pkg/ring/kv/codec"
-	"github.com/cortexproject/cortex/pkg/ring/kv/memberlist"
 	"github.com/cortexproject/cortex/pkg/util/log"
 	"github.com/go-kit/kit/log/level"
+	"github.com/grafana/dskit/kv/codec"
+	"github.com/grafana/dskit/kv/memberlist"
 	"github.com/grafana/dskit/modules"
 	"github.com/grafana/dskit/services"
 	"github.com/prometheus/client_golang/prometheus"
+	"github.com/thanos-io/thanos/pkg/discovery/dns"
 	"github.com/weaveworks/common/middleware"
 	"github.com/weaveworks/common/server"
 
@@ -101,6 +102,8 @@ func (t *App) initOverrides() (services.Service, error) {
 		return nil, fmt.Errorf("failed to create overrides %w", err)
 	}
 	t.overrides = overrides
+
+	t.Server.HTTP.Handle("/runtime_config", overrides.Handler())
 
 	return t.overrides, nil
 }
@@ -248,21 +251,31 @@ func (t *App) initStore() (services.Service, error) {
 }
 
 func (t *App) initMemberlistKV() (services.Service, error) {
-	t.cfg.MemberlistKV.MetricsRegisterer = prometheus.DefaultRegisterer
+	reg := prometheus.DefaultRegisterer
+	t.cfg.MemberlistKV.MetricsRegisterer = reg
 	t.cfg.MemberlistKV.MetricsNamespace = metricsNamespace
 	t.cfg.MemberlistKV.Codecs = []codec.Codec{
 		ring.GetCodec(),
 	}
 
-	t.memberlistKV = memberlist.NewKVInitService(&t.cfg.MemberlistKV, log.Logger)
+	dnsProviderReg := prometheus.WrapRegistererWithPrefix(
+		"tempo_",
+		prometheus.WrapRegistererWith(
+			prometheus.Labels{"name": "memberlist"},
+			reg,
+		),
+	)
 
-	t.cfg.Ingester.LifecyclerConfig.RingConfig.KVStore.MemberlistKV = t.memberlistKV.GetMemberlistKV
-	t.cfg.Distributor.DistributorRing.KVStore.MemberlistKV = t.memberlistKV.GetMemberlistKV
-	t.cfg.Compactor.ShardingRing.KVStore.MemberlistKV = t.memberlistKV.GetMemberlistKV
+	dnsProvider := dns.NewProvider(log.Logger, dnsProviderReg, dns.GolangResolverType)
+	t.MemberlistKV = memberlist.NewKVInitService(&t.cfg.MemberlistKV, log.Logger, dnsProvider, reg)
 
-	t.Server.HTTP.Handle("/memberlist", t.memberlistKV)
+	t.cfg.Ingester.LifecyclerConfig.RingConfig.KVStore.MemberlistKV = t.MemberlistKV.GetMemberlistKV
+	t.cfg.Distributor.DistributorRing.KVStore.MemberlistKV = t.MemberlistKV.GetMemberlistKV
+	t.cfg.Compactor.ShardingRing.KVStore.MemberlistKV = t.MemberlistKV.GetMemberlistKV
 
-	return t.memberlistKV, nil
+	t.Server.HTTP.Handle("/memberlist", t.MemberlistKV)
+
+	return t.MemberlistKV, nil
 }
 
 func (t *App) setupModuleManager() error {
@@ -282,8 +295,8 @@ func (t *App) setupModuleManager() error {
 
 	deps := map[string][]string{
 		// Server:       nil,
-		// Overrides:    nil,
 		// Store:        nil,
+		Overrides:     {Server},
 		MemberlistKV:  {Server},
 		QueryFrontend: {Server},
 		Ring:          {Server, MemberlistKV},