+++ tempodb/encoding/common/types.go
@@ -24,10 +24,10 @@ type Record struct {
 
 // ObjectCombiner is used to combine two objects in the backend
 type ObjectCombiner interface {
-	// Combine objA and objB encoded using dataEncoding. The returned object must
+	// Combine objects encoded using dataEncoding. The returned object must
 	// use the same dataEncoding. Returns a bool indicating if it the objects required combining and
 	// the combined slice
-	Combine(objA []byte, objB []byte, dataEncoding string) ([]byte, bool)
+	Combine(dataEncoding string, objs ...[]byte) ([]byte, bool)
 }
 
 // DataReader returns a slice of pages in the encoding/v0 format referenced by